refactor(chat-dataflow): extract _chat_pre_safety helper from chat_node
